+++ test/function/arithmetic/pow.test.js
@@ -244,7 +244,7 @@ describe('pow', function() {
   });
 
   it('should compute large size of square matrix', function() {
-    var a = math.eye(30).valueOf();
+    var a = math.identity(30).valueOf();
     approx.deepEqual(pow(a, 1000), a);
     approx.deepEqual(pow(matrix(a), 1000), matrix(a));
   });